+++ src/heuristics.h
@@ -6,8 +6,8 @@
 #define FS_PROTOTYPE_HEURISTICS_H
 
 #include <vector>
-int HammingDistance(int size, std::vector<int> board);
+double HammingDistance(int size, std::vector<int> board);
 double CartesianDistance(int size, std::vector<int> board);
-int ManhattanDistance(int size, std::vector<int> board);
+double ManhattanDistance(int size, std::vector<int> board);
 
 #endif //FS_PROTOTYPE_HEURISTICS_H
